requirements.txt
streamlit==1.25.0
gtts==2.3.2
speechrecognition==3.8.1
torch==2.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyPDF2
huggingface_hub==0.14.1
transformers==4.30.0
speechrecognition
openai==0.27.8
python-dotenv